Narrow the per-glyph feature masks to 32 bits
The global-bit classification bounds distinct bits at roughly eleven
per plan, so the three per-glyph masks drop from 8 to 4 bytes each,
removing 12 bytes from the glyph record. The shared global bit moves
from bit 63 to bit 31 and 28 assignable varying bits remain above the
reserved vertical trio.

Benchmark (12 iterations, clean control pair, measured with the ASCII
grapheme fast path also in tree; ratios are the load-independent
metric):

| Scenario   | Before           | After            | Allocated |
|------------|------------------|------------------|-----------|
| Latin      | 5.69 us / 2.49x  | 5.06 us / 2.23x  | 6,216 B   |
| Arabic     | 7.57 us / 1.93x  | 7.62 us / 1.94x  | 4,368 B   |
| Devanagari | 12.05 us / 1.30x | 11.74 us / 1.32x | 3,832 B   |

The Latin gain carries the ASCII fast path committed separately; this
change is judged on non-regression, and Arabic and Devanagari ratios
sit within measurement noise of their baselines. Allocations are
byte-identical. Suite 5,585/0 including the differential oracle.

+++ src/SixLabors.Fonts/GlyphShapingData.cs
@@ -245,14 +245,14 @@ internal struct GlyphShapingData
     /// only ever reveals a registered bit; a feature that was never registered for the
     /// glyph cannot be enabled.
     /// </summary>
-    public ulong RegisteredFeatureMask { get; set; }
+    public uint RegisteredFeatureMask { get; set; }
 
     /// <summary>
     /// Gets or sets the mask of features currently enabled for this glyph: the subset of
     /// <see cref="RegisteredFeatureMask"/> a lookup application gate tests with a single
     /// bitwise AND.
     /// </summary>
-    public ulong FeatureMask { get; set; }
+    public uint FeatureMask { get; set; }
 
     /// <summary>
     /// Gets or sets the mask of features whose lookups actually changed this glyph.
@@ -261,7 +261,7 @@ internal struct GlyphShapingData
     /// therefore survives the copy into the positioning buffer regardless of which
     /// plan wrote it.
     /// </summary>
-    public ulong AppliedFeatureMask { get; set; }
+    public uint AppliedFeatureMask { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether this glyph is the result of a substitution.
