+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupType4SubTable.cs
@@ -154,7 +154,7 @@ internal sealed class LookupType4Format1SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
