+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupType5SubTable.cs
@@ -88,7 +88,7 @@ internal sealed class LookupType5Format1SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
@@ -243,7 +243,7 @@ internal sealed class LookupType5Format2SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
@@ -415,7 +415,7 @@ internal sealed class LookupType5Format3SubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
     {
