+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/NotImplementedSubTable.cs
@@ -23,7 +23,7 @@ internal class NotImplementedSubTable : LookupSubTable
         GSubTable table,
         ShapingBuffer buffer,
         Tag feature,
-        ulong lookupMask,
+        uint lookupMask,
         int index,
         int count)
         => false;
